Backend/Travel/ai-smart-travel-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.11.3
python-dotenv==1.1.0
pytest==8.3.5
